fix(tray): separate qt and maocOS test cases
