+++ test/CMakeLists.txt
@@ -102,7 +102,11 @@ nextcloud_add_test(IconUtils)
 nextcloud_add_test(NotificationSoundPlayer)
 nextcloud_add_test(SetUserStatusDialog)
 nextcloud_add_test(TrayAccountMenuPolicy)
-nextcloud_add_test(SystraySyncControl)
+if(APPLE)
+    nextcloud_add_test(SystraySyncControlMacOS)
+else()
+    nextcloud_add_test(SystraySyncControlQt)
+endif()
 nextcloud_add_test(UnifiedSearchListmodel)
 nextcloud_add_test(ActivityListModel)
 nextcloud_add_test(SortedActivityListModel)
